Move payment-reference-calculator.ts
