+++ packages/request-client.js/src/api/payment-network/eth/payment-reference-calculator.ts
